% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debate_contributions.R
\name{get_debate_contributions}
\alias{get_debate_contributions}
\title{Get debate contributions}
\usage{
get_debate_contributions(access_token = access_token)
}
\value{
A data frame of debate contributions and corresponding meta data.
}
\description{
\code{get_debate_contributions()} gets the debate data from EuroParlBase.
}
