% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mep_political.R
\name{get_mep_political}
\alias{get_mep_political}
\title{Get political MEP data}
\usage{
get_mep_political(access_token = NULL)
}
\description{
\code{get_mep_political()} gets the political MEP data from EuroParlBase.
}
