% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sittings.R
\name{get_sittings}
\alias{get_sittings}
\title{Get plenary sitting data}
\usage{
get_sittings(access_token = access_token)
}
\description{
\code{get_sittings()} gets the plenary sittings data from EuroParlBase.
}
