% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vote_infos.R
\name{get_vote_infos}
\alias{get_vote_infos}
\title{Get information on roll-call votes}
\usage{
get_vote_infos(access_token = NULL)
}
\description{
\code{get_vote_infos()} gets information on the roll-call votes from EuroParlBase.
}
