% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_votes.R
\name{get_votes}
\alias{get_votes}
\title{Get individual roll-call voting decisions}
\usage{
get_votes(access_token = NULL, term = NULL)
}
\description{
\code{get_votes()} gets MEPs' roll-call voting decisions from EuroParlBase.
}
