% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_debate_contributions.R
\name{get_debate_contributions}
\alias{get_debate_contributions}
\title{Get debate contributions}
\usage{
get_debate_contributions(access_token = NULL, corpus = "original")
}
\description{
\code{get_debate_contributions()} gets debate contributions from EuroParlBase.
}
